Add ComponentBaseWithState
