+++ src/MudBlazor/Base/MudComponentBase.cs
@@ -1,20 +1,19 @@
-﻿using System;
+﻿// Copyright (c) MudBlazor 2021
+// MudBlazor licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
 using System.Collections.Generic;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.Logging;
 using Microsoft.JSInterop;
 using MudBlazor.Interfaces;
-using MudBlazor.State;
-using MudBlazor.State.Builder;
 
 namespace MudBlazor
 {
 #nullable enable
-    public abstract partial class MudComponentBase : ComponentBase, IMudStateHasChanged
+    public abstract class MudComponentBase : ComponentBaseWithState, IMudStateHasChanged
     {
-        internal readonly ParameterSet Parameters;
-
         [Inject]
         private ILoggerFactory LoggerFactory { get; set; } = null!;
         private ILogger? _logger;
@@ -61,37 +60,11 @@ namespace MudBlazor
                     ? (id.ToString() ?? $"mudinput-{Guid.NewGuid()}")
                     : $"mudinput-{Guid.NewGuid()}";
 
-        protected MudComponentBase()
-        {
-            Parameters = new ParameterSet(this);
-            _scope = new ParameterRegistrationBuilderScope(this);
-        }
-
         /// <inheritdoc />
         protected override void OnAfterRender(bool firstRender)
         {
             IsJSRuntimeAvailable = true;
             base.OnAfterRender(firstRender);
-        }
-
-        /// <inheritdoc />
-        protected override void OnInitialized()
-        {
-            base.OnInitialized();
-            Parameters.OnInitialized();
-        }
-
-        /// <inheritdoc />
-        public override Task SetParametersAsync(ParameterView parameters)
-        {
-            return Parameters.SetParametersAsync(base.SetParametersAsync, parameters);
-        }
-
-        /// <inheritdoc />
-        protected override void OnParametersSet()
-        {
-            base.OnParametersSet();
-            Parameters.OnParametersSet();
         }
 
         /// <inheritdoc />
