+++ src/MudBlazor/Extensions/ComponentBaseWithStateExtensions.cs
@@ -1,4 +1,8 @@
-﻿using System;
+﻿// Copyright (c) MudBlazor 2021
+// MudBlazor licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using MudBlazor.State;
@@ -7,9 +11,9 @@ namespace MudBlazor.Extensions;
 
 #nullable enable
 /// <summary>
-/// Provides extension methods for MudBlazor components to facilitate accessing parameter states.
+/// Provides extension methods for <see cref="ComponentBaseWithState"/> components to facilitate accessing parameter states.
 /// </summary>
-internal static class MudComponentExtensions
+internal static class ComponentBaseWithStateExtensions
 {
     /// <summary>
     /// Gets the read-only parameter state of a specified property in the component.
@@ -23,7 +27,7 @@ internal static class MudComponentExtensions
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="propertyExpression"/> is null.</exception>
     /// <exception cref="ArgumentException">Thrown when <paramref name="propertyExpression"/> is invalid.</exception>
     /// <exception cref="KeyNotFoundException">Thrown when the parameter state with <paramref name="propertyExpression"/> is not found.</exception>
-    public static T? GetState<TComponent, T>(this TComponent component, Func<TComponent, T> propertyExpression, [CallerArgumentExpression(nameof(propertyExpression))] string? propertyNameCallerArgumentExpression = null) where TComponent : MudComponentBase
+    public static T? GetState<TComponent, T>(this TComponent component, Func<TComponent, T> propertyExpression, [CallerArgumentExpression(nameof(propertyExpression))] string? propertyNameCallerArgumentExpression = null) where TComponent : ComponentBaseWithState
     {
         ArgumentNullException.ThrowIfNull(propertyExpression);
         var propertyName = GetPropertyFromFuncLambda(propertyNameCallerArgumentExpression);
@@ -39,7 +43,7 @@ internal static class MudComponentExtensions
     /// <param name="propertyName">The name of the property whose parameter state needs to be accessed. Use nameof(...) to get the property name.</param>
     /// <returns>The <see cref="ParameterState{T}.Value"/> of the specified property.</returns>
     /// <exception cref="KeyNotFoundException">Thrown when the parameter state with <paramref name="propertyName"/> is not found.</exception>
-    public static T? GetState<T>(this MudComponentBase component, string propertyName)
+    public static T? GetState<T>(this ComponentBaseWithState component, string propertyName)
     {
         if (component.Parameters.TryGetValue(propertyName, out var lifeCycle))
         {
